Pricing updates (#8818)
* use test

* addon defaults
